requirements.txt
google-api-python-client
langchain
langchain_chroma
langchain-google-genai
google-generativeai
python-dotenv
flask
flask-cors
pandas
nltk
streamlit
colorama
protobuf==3.20.3
plotly
matplotlib
wordcloud
unidecode 
langdetect 
emoji
langchain_core
langchain_community
torch
transformers
ml-dtypes
#pysqlite3-binary

chromadb==0.4.24
numpy==1.26.4
posthog==5.0.0
tqdm
gunicorn
watchdog
langchain-ollama